requirements.txt
openai
pyttsx3
Flask
requests
schedule
sounddevice
numpy
scipy
openai-whisper